+++ src/literalizer/languages/forth.py
@@ -15,6 +15,7 @@ from literalizer._formatters.collection_openers import (
 )
 from literalizer._formatters.format_dates import (
     format_datetime_epoch,
+    format_time_iso,
 )
 from literalizer._formatters.format_entries import (
     assignment_formatter_from_declaration,
@@ -701,6 +702,11 @@ class Forth(metaclass=LanguageCls):
     def format_datetime(self) -> Callable[[datetime.datetime], str]:
         """Callable that formats a datetime as a string literal."""
         return self.datetime_format
+
+    @cached_property
+    def format_time(self) -> Callable[[datetime.time], str]:
+        """Callable that formats a time as a string literal."""
+        return format_time_iso
 
     @cached_property
     def format_string(self) -> Callable[[str], str]: